Mild refactor in preparation
