+++ libs/storage/Tsavorite/cs/src/core/Index/Synchronization/HybridLogCheckpointStateMachine.cs
@@ -1,0 +1,53 @@
+﻿// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
+
+namespace Tsavorite.core
+{
+    /// <summary>
+    /// Hybrid log checkpoint state machine.
+    /// </summary>
+    internal class HybridLogCheckpointStateMachine<TKey, TValue, TStoreFunctions, TAllocator> : VersionChangeStateMachine<TKey, TValue, TStoreFunctions, TAllocator>
+        where TStoreFunctions : IStoreFunctions<TKey, TValue>
+        where TAllocator : IAllocator<TKey, TValue, TStoreFunctions>
+    {
+        /// <summary>
+        /// Construct a new HybridLogCheckpointStateMachine to use the given checkpoint backend (either fold-over or
+        /// snapshot), drawing boundary at targetVersion.
+        /// </summary>
+        /// <param name="checkpointBackend">A task that encapsulates the logic to persist the checkpoint</param>
+        /// <param name="targetVersion">upper limit (inclusive) of the version included</param>
+        public HybridLogCheckpointStateMachine(ISynchronizationTask<TKey, TValue, TStoreFunctions, TAllocator> checkpointBackend, long targetVersion = -1)
+            : base(targetVersion, new VersionChangeTask<TKey, TValue, TStoreFunctions, TAllocator>(), checkpointBackend) { }
+
+        /// <summary>
+        /// Construct a new HybridLogCheckpointStateMachine with the given tasks. Does not load any tasks by default.
+        /// </summary>
+        /// <param name="targetVersion">upper limit (inclusive) of the version included</param>
+        /// <param name="tasks">The tasks to load onto the state machine</param>
+        protected HybridLogCheckpointStateMachine(long targetVersion, params ISynchronizationTask<TKey, TValue, TStoreFunctions, TAllocator>[] tasks)
+            : base(targetVersion, tasks) { }
+
+        /// <inheritdoc />
+        public override SystemState NextState(SystemState start)
+        {
+            var result = SystemState.Copy(ref start);
+            switch (start.Phase)
+            {
+                case Phase.IN_PROGRESS:
+                    result.Phase = Phase.WAIT_FLUSH;
+                    break;
+                case Phase.WAIT_FLUSH:
+                    result.Phase = Phase.PERSISTENCE_CALLBACK;
+                    break;
+                case Phase.PERSISTENCE_CALLBACK:
+                    result.Phase = Phase.REST;
+                    break;
+                default:
+                    result = base.NextState(start);
+                    break;
+            }
+
+            return result;
+        }
+    }
+}
